requirements.txt
django
django-money
django-paypal
pillow
python-decouple
requests
cryptography
